[FEATURE] Add option to force a redirect
Add an option to force a redirect if the detected
language is the default language of the installation.

Forcing a redirect may be reasonable if an extension
like RealUL is used and the default language shall be
show in the url
(see https://forge.typo3.org/issues/43772).

New TypoScript parameter:
* »plugin.tx_rlmplanguagedetection_pi1.forceRedirect« (bool, default = 0),