[Feature] Allow language redirect & deep linking
Let's take a website with protected pages and external links pointing to them :
Login page (not access restricted)
- Protected page 1
- Protected page 2
- etc
Having left [FE][pageNotFound_handling] empty, the default behaviour when trying to access directly "Protected page 1" is to show the login page, with the referrer left intact. This works well without rlmp_language_detection, but it doesn't with the plugin enabled and no language specified, because the language redirect uses the id of the login page, instead of the targeted page.

The proposed change fixes this issue, for Typo3 v7+ (the method "getRequestedId" doesn't exist before). On v6.x, the current behaviour is kept.